requirements.txt
gunicorn
django
djangorestframework-datatables-editor
djangorestframework